Move ColorListItem to its own file
